[SYCL][E2E] Fix error reporting in e2e lit config
Checking the output of `get_sycl_ls_verbose` expected the function to
update a global `sp` variable. Since it didn't do that, this resulted
in an error when trying to print the command output.

This patch changes the function to return the subprocess directly, which
is used during error message generation.