[SYCL][E2E][CI] Disable `std_array.cpp` at CI level
The issue is specific to the post-commit CI using OneAPI compiler to
build SYCL RT (see https://github.com/intel/llvm/issues/18458). Ideally,
we'd like to use `/Qno-intel-lib` during such build but OneAPI seems to
have a bug related to that (reported internally). It will take an
official release for it to reach this CI even after it's implemented, so
we need to disable the test on that specific configuration only for the
time being.