[SYCL][NFC] Add E2E test for gpu AOT to JIT fallback (#18492)
I believe the feature of AOT to JIT fallback was implemented and enabled
in this commit
https://github.com/intel/llvm/commit/38e588da39f7492570ea66c10f0fcc465562bddf

commit includes UT for jit fallback for kernel compiled for older
platform
https://github.com/intel/llvm/commit/38e588da39f7492570ea66c10f0fcc465562bddf

this fallback works when `-fsycl-targets` is set for compilation and AOT
images contain info about architecture it is built for. In order to
build program SYCL RT choose images by compile target (if property is
set) and adds spirv (if available) to the image set passed to backend
later.

Choice what image to use is done in UR adapter for specific backend.
Where it chooses AOT image if available and uses SPIRV as fallback.
While RT part is completely tested by UT, adding small E2E test to
verify stack behavior.

---------

Signed-off-by: Tikhomirova, Kseniya <kseniya.tikhomirova@intel.com>
Co-authored-by: Marcos Maronas <marcos.maronas@intel.com>