[libspirv] Add missing dependency; delete unused file (#18569)
The 'common' conversion python script is no longer really that common,
only being used by the SPIR-V conversion script. This commit moves it
next to its only user, and adds a missing dependency on it in CMake.
This should ensure a rebuild is triggered when that file is modified.

Additionally, this commit deletes the unused 'gen_core_convert.py' since
upstream uses regular (non-libspirv) gen_convert.py with the '--clc'
flag.