[libspirv] Avoid duplicating builtins during remangling (#18559)
This commit adds some checks to the remangler. It prevents it from
generating clones or renaming functions if a function with the target
name already exists in the module.

This can happen if the libspirv source provides, e.g., both a char and a
signed char version of a builtin and the remangler wants to remangle
char to signed char.

In such cases we'd previously generate new unreachable builtins as to
avoid the naming clash LLVM would give them a random suffix like
renaming _Z3fooi to _Z3fooi24234. We can't know that the functions are
equivalent but this situtation is no worse than before - we'd be calling
into the 'wrong' builtin in any case.

This removes around 2000 unreachable builtins from the 'l64' remangled
libraries and 300 from the 'l32' ones.

This commit does introduce some new builtins too. Some builtins like
_Z17__spirv_ocl_s_minDv16_cS_ are now in the 'unsigned char' module
which I believe is okay; any user calling 'char' builtins on an
char-is-unsigned-char platform would be calling into the unsigned char
versions ('h') anyway. These builtins are now included because they are
in the source module and left alone when the remangler detects that the
remangled version already exists.