[UR][Graph] Disable flaky PVC CTS test
CTS test `urInOrderUSMCommandBufferExpTest.WithHints` has been reported
as failing on unrelated PRs on PVC in GitHub CI https://github.com/intel/llvm/issues/18544

Disable the UR CTS in-order tests on Level-Zero V1 adapter until this
bug is investigated and resolved. Disabled more than the specific test
that fails, as the others are variants of it, and it is not unlikely
that we'd end up seeing flaky fails in CI for them too.