[UR][Graph] Disable flaky PVC CTS test (#18586)
CTS test `urInOrderUSMCommandBufferExpTest.WithHints` has been reported
as failing on unrelated PRs on PVC in GitHub CI
https://github.com/intel/llvm/issues/18544 which were added recently in
#18444

Disable the UR CTS in-order tests on Level-Zero V1 adapter until this
bug is investigated and resolved. I've disabled more than the specific
test that fails, as the others UR tests inheriting from the base fixture
are variants, and it is not unlikely that we'd end up seeing flaky fails
in CI for them too.