[libspirv] Remove unnecessary SPIR-V builtin decls
We rely on clang's built in knowledge of certain SPIR-V builtins and its
ability to generate declarations for them on the fly, namely
conversions. This means we have over 10,000 lines of unnecessary code
that's difficult to maintain.

Even if we did want to provide our own declarations, libclc provides
more concise ways of expressing declarations using macros which we could
use instead.