[SYCL][Bindless][E2E] Match UUID when searching for Vulkan device (#18562)
Currently the Vulkan interop tests create a Vulkan device based on the
name of the SYCL device. This only works for very simple cases where the
Vulkan driver
and the SYCL backend driver report a similar enough name.

With this patch, a Vulkan device is created from a SYCL one by matching
the device UUID, which guarantees the same device.