[CI][Benchmark] Switch off L0 for computeBench on SYCL for CUDA (#18487)
Currently L0/opencl benchmarks is still built on computeBench even when
we specify that we need to benchmark sycl on top of cuda that creates an
error sometimes when the execution environment doesn't have L0 setup.
That PR will close building L0/opencl benchmarks on computeBench when we
specify cuda target for benchmarking.