[libclc] Ensure CLC builtins see the generic addrspace (#18585)
Having the macros defined in either the SPIR-V headers or OpenCL headers
left it easy for the generic address space macros to be undefined.
What's worse is there was no warning or error that this was happening.

This commit moves the definition of these macros into a common shared
header that all three builtin libraries see. It also ensures the build
will fail on the use of any undefined macro. This mirrors how upstream
will (likely) eventually place the code, though in all likelihood it
will differ slightly.