[SYCL] the first kernel function declaration should be added with attribute (#18405)
Docs [Defining a free function
kernel](https://github.com/intel/llvm/blob/sycl/sycl/doc/extensions/proposed/sycl_ext_oneapi_free_function_kernels.asciidoc#defining-a-free-function-kernel)
contains:
The property must appear on the first declaration of the function in the
translation unit. Redeclarations of the function may optionally be
decorated with the same property if the property argument is the same.
The effect is the same regardless of whether redeclarations are so
decorated.

This PR changes `isFreeFunction` to define if the first occurrence of
kernel free function has attribute or not. New clang diagnostics was
added to highlight that first occurrence does not have attribute but the
next has.