[libclc] Align generic AS support with upstream (#18617)
After a merge resolution we were left with two essentially equivalent
methods of enabling generic address space support in libclc and
libspirv. This commit removes the old downstream implementation.

There was also a merge conflict resolution bug in that we were
multiply-defining a macro, which could lead to errors when compiling.

One downstream-specific libspirv builtin was unconditionally assuming
generic AS support, which we don't assume upstream. This presumably
isn't a problem in our regular SYCL builds, but other upstream libclc
targets fail when hitting this code.