[libclc] Reduce diff with upstream in CLC headers (#18625)
In a recent sycl-web PR I mistakenly introduced extra includes to some
CLC functions to account for the generic address space. This was because
the macro defines weren't at that time being pulled in by
math/gentype.inc (or any CLC headers) as they were in my upstream work.

After pulling in the upstream generic address space support, the defines
are now in a common place across the whole of libclc and are
automatically included by all of these files. Keeping the redundant
includes around only serves to increases our diff with upstream.

This commit takes the opportunity to clean up some whitespace diffs at
the same time.