[UR][CUDA][HIP] Refactor event handling
- Use constructor instead of unnecessary makeNative function
- Move profiling stream constructor to queue and timestamp command
- Simplify even creation code
- Use make_unique
- Re-organize event.hpp and event.cpp

This is mostly NFC, the one behavior change is that we no longer create
the profiling stream in event creation, it will be created either when
we build a profiling queue, or when we enqueue a timestamp event. Which
mostly means that it will be created a bit sooner for profiling queues.