[SYCL][E2E] Better handle the Intel Compiler in standalone build (NFC)
The proprietary Intel Compiler (ICX) uses a different installation
layout then clang. It puts tools into a bin/compiler subdirectory
(to not expose them on the PATH by default). Handle this by not assuming
that the compiler is in the same directory as other llvm tools.
Ask the compiler for the path to "llvm-config" for the tools directory,
using the "-print-prog-name" option.
llvm-config was choosen because lit already assumes it is in the tools
directory see for example: [llvm/utils/lit/lit/llvm/config.py:285][1]

[1]: https://github.com/llvm/llvm-project/blob/9cac4bf485e64f7992f2c01bb9517f6379e58164/llvm/utils/lit/lit/llvm/config.py#L285