[libspirv] Make OpenCL not depend on SPIR-V
This commit finally frees the OpenCL builtins from their dependency on
any SPIR-V dependencies.

The last thing was a header in the OpenCL directory which was only used
by libspirv.  This commit moves that to a more logical place. We can
then officially remove the dependencies.

A mistaken conflict resolution with regards to disabling an error has
also been fixed up.