[SYCL][E2E] Fix `XFAIL` statements that use negated device features (#18641)
Because of device specific features we have to perform our own checks to
see whether or not a test should XFAIL. However, a similar check is also
done in internal lit code, but due to the fact that this bit of code
only knows about the non device specific features there can be
situations where it incorrectly reports a test as expected to fail (Like
negating a device specific feature, like `XFAIL: !cpu`).

To avoid this we set the XFAIL conditions to an empty list before
returning our result, this way `expected_to_fail` is always false in the
code below, and our result is not altered.


https://github.com/intel/llvm/blob/d60cd271ccb330dd9da8158176404b68ff55baa9/llvm/utils/lit/lit/Test.py#L275-L292