Only build libspirv for selected targets
In practice we only build (and thus test) three libclc targets:
'nvptx64--nvidiacl', 'amdgcn--amdhsa' and 'native_cpu'. All other
upstream libclc targets are never built in our CI and would in fail to
build.

This commit rectifies this by selectively building libspirv only for
those three supported targets. More can be added in time if required.

There are still certain OpenCL libclc targets that can't be built with
this commit. The r600 target, for example, can't build because we
unconditionally enable the fp64 OpenCL extension across the board, but
the r600 target doesn't support that.

The clspv and clspv64 targets also fail to build due to SOURCES files
referencing missing files. This will be resolved in the next pulldown.