[UR] Fix logger use after move (Coverity)
When the callback sink and the standard sink are both enabled using
`std::forward<Args>(args)...` results in a use-after-move. This patches
fixes the issue by letting the r-value references decay to l-value
references when expanding the `args...` parameter pack and allowing the
parameter pack to be passed to multiple functions. The `logger::log()`
member function does not need to take ownership of the objects being
passed in.