[SYCL][CUDA] Expand search for generated_cuda_meta.h (#18436)
On some systems `CUDA_Toolkit_ROOT` might be emtpy,
even though CUDA is properly found in the CMake.
This can cause a failed search for `generated_cuda_meta.h`.
In that case only a warning is emitted when building
`cuda_trace_collector`,
but this can fail the `-Werror` build.
This patch ensures that if we can find `CUPTI`, we can find this file.