[libspirv] Avoid macro redefinition
The macro __CLC_FUNCTION is special and is used in CLC headers. Defining
it here in an implementation file - nextafter.cl - for another purpose
then including half_nextafter.inc which pulls in the SPIR-V headers
results in a macro redefinition, which is warned about.

Sicne the name isn't important and is local to this one file, this
commit just changes the macro definition to fix the issue.