[SYCL][NFC] Avoid unused variable AccTarget (#18675)
After https://github.com/intel/llvm/pull/18437, the runtime library is
producing a warning about an unused variable AccTarget in handler.cpp.
This is due to the variable only being used in assert, which may in turn
be removed when assertions are disabled. This commit removes the
variable in favor of making the conversion inside the assert.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>