[SYCL][E2E] Check for run-time features when in build-only (#17988)
Currently when we check `REQUIRES`/`UNSUPPORTED` statements in
`build-only` mode we throw an error if any feature that is not in
`build_specific_features` evaluates to true. Since this is error will
only trigger when a test queries for one of these features, it can be
easy to miss adding new build-specific features to this list (see
#17985, and #17363).

This pr changes this check to be done for all features when in
`build-only`, not just those queried by a test. If any available feature
does not appear in `build_specific_features` then an error is thrown.