[SYCL][NVPTX][AMDGCN] Move devicelib cmath to header
This patch experiments with moving standard library math built-ins from
libdevice into headers.

This is based on the way clang handles this for CUDA and HIP. In these
languages you can define device functions as overloads. This allows
re-defining standard library functions specifically for the device in a
header, so that we can provide a device specific implementations of
certain built-ins while still using the regular standard library
headers.

By default SYCL doesn't do overloads for device functions, so this patch
introduces a new `sycl_device_only` attribute, this attribute will make
a function device only and allow it to overload with existing functions.