[CI] Test 'all' cmake target in nightly (#18701)
There was a case where we broke build for the `all` target but
`sycl-toolchain` worked, so build `all` to catch this.

Nightly run
[here](https://github.com/intel/llvm/actions/runs/15302502386), failures
not because of this.

We need the `||` part in ` ${{ inputs.build_target || 'sycl-toolchain'
}}` because for the `workflow_dispatch` case, we already have 10 input
args which is the max so we can't always have a `inputs.build_target`
available without more work.

Closes: https://github.com/intel/llvm/issues/18648

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>
Co-authored-by: aelovikov-intel <andrei.elovikov@intel.com>