[SYCL] Update time bases when submit time exceeds start time
Currently for command buffers there is a workaround to set submission time
to start time if the recorded submit time is greater than the start time.
This is needed because the submission time is based on estimated clock
and over time device time and host time may drift apart causing
aforementioned estimation to be inaccurate.

I belive same "workaround" should be used for all tasks, not only for
command buffers.

More over it seems like this condition can be used as a trigger to query
synchronized clock from backend, i.e. to correct the estimation. This PR
implements that.