[SYCL] fix for std::locale::global interfering with allowlist check (#18550)
`std::locale::global()` is a very common way for apps to initialize
string localization for numeric and monetary values. Unfortunately, it
is global.
In this very limited fix, we are preventing it from interfering with the
allowlist check which uses `std::stringstream`. The insertion of commas
into numbers for the locale support messes with the expected values.
We'll probably have to audit the other uses of std streams to see if
they would be similarly vulnerable.

---------

Signed-off-by: Chris Perkins <chris.perkins@intel.com>