[SYCL][UR] Fix the accuracy of command submission timestamp
Current submission time calculation is inaccurate as it is not using
synchronized timestamps returned by zeDeviceGetGlobalTimestamps but
using only device timestamp from that call and uses std::chrono close to
that call to record the host time. This estimation becomes inaccurate
pretty quickly, for example, on 5000 iterations for the test that we
have. This PR fixes this problem using the fact that L0 runtime
implementation uses CLOCK_MONOTONIC_RAW on Linux and
QueryPerformanceCounter on Windows.