[SYCL] Fix use-after-free for MockHandler::getKernelName()
In current implementation, std::string is returned. For preview, caller
of getKernelName() creates std::string_view from the temporary string.
This sting_view is used after the temporary string is destroyed. In the
fix for preview std::string_view is returned, that points to same kernel
name as handler::MKernel.