[SYCL] Add "ext_oneapi_offload" backend enum value (#18683)
This adds `offload` as a backend to the sycl runtime. This means it is
accepted as a value for `ONEAPI_DEVICE_SELECTOR` and friends, and
prevents an offload device reporting itself as such causing an exception
to be thrown.

Note that the UR offload adapter is still not built as part of the
build. The only way to use this is with a standalone build of
unified-runtime with libur_adapter_offload enabled and in the
`LD_LIBRARY_PATH`.