[SYCL] Pass `handler &` instead of queue across ABI for reduction utils
Queue might be `nullptr` in case of graph, but the information this
utils query is device-specific. By passing entire `handler &` and having
access to graph information we'd be able to return more precise results.

Another positive side-effect is that we eliminiate explicit
`std::shared_ptr<queue_impl>` which is a small step forward in the
ongoing refactoring efforts to prefer passing `*_impl` by raw ptr/ref
with explicit `shared_from_this` whenever lifetimes need to be extended.