[SYCL][UR] Include backend name in exception messages
To aid with more accurate identification of the source of an error in
the event of a thrown exception, include the backend name in error
messages resulting from calls to Unified Runtime entry points. Here's an
example of the output.

Before:

```
==== DeviceSanitizer: ASAN
terminate called after throwing an instance of 'sycl::_V1::exception'
 what(): UR backend failed. UR backend returns:40 (UR_RESULT_ERROR_OUT_OF_RESOURCES)
```

After:

```
==== DeviceSanitizer: ASAN
terminate called after throwing an instance of 'sycl::_V1::exception'
 what(): level_zero backend failed with error: 40 (UR_RESULT_ERROR_OUT_OF_RESOURCES)
```