[SYCL] Make host task timestamps share the same base as device tasks (#18710)
Currently, we use `std::chrono` to record `command_submit`,
`command_start` and `command_end` timestamps for host tasks. So, if
there there is a mix of device tasks and host tasks submitted to the
queue, they have different time base and inconvenient to use/compare.
This PR makes host task timestamps to have the same time base as device
tasks.