[UR] Ensure mock target depends on UR_EXTERNAL_DEPENDENCIES (#18844)
Firstly, rename `ur_mock_headers` to `ur_mock` as this is a shared
library not header only library.

Secondly, when part of a subbuild, ensure that the `ur_mock` target
depends on the specified `UR_EXTERNAL_DEPENDENCIES`, this is handled by
using `add_ur_library()` in place of `add_library()`.