[SYCL][E2E][BINDLESS] Add ptr to ptr of image_handle test failing on level zero only (#18721)
This test passes on the cuda backend but fails on level_zero. It
allocates a single element array of `unsampled_image_handle*`, pointing
to a valid `unsampled_image_handle` that is dereferenced on the device.
The casting to generic `void*` is not necessary to make the test fail on
l0, but is done to keep this test matching the original bug report we
received.

Note that rewriting this test to make an identical dereference with for
example a single 2d C array of `long` passes on level_zero, indicating
that this issue is specific to image_handles.

---------

Signed-off-by: JackAKirk <jack.kirk@codeplay.com>