[SYCL][NFC] Remove PropertyValue::copy() (#18945)
This commit removes the PropertyValue::copy() foot-gun. This function
simply copies the values from one PropertyValue to the other, but for
byte arrays, this means the caller of the copy need to ensure that the
underlying memory isn't managed by both the original and the copy.
Removing the function makes it clearer at the call-sites that this
ownership relation has to be maintained explicitly.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>