[CI][Benchmarks] fix LD_LIBRARY_PATH set for compute runtime
The LD_LIBRARY_PATH for the level-zero loader and gmmlib, which are
dependencies of compute runtime, were set incorrectly to lib64.
This led to issues compatibility issues when running benchmarks.

This patch sets the correct path and adds a warning message so that
it's easier to notice such issues in the future.