[CI][Benchmarks] fix LD_LIBRARY_PATH set for compute runtime (#18951)
The LD_LIBRARY_PATH for the level-zero loader and gmmlib, which are
dependencies of compute runtime, was set incorrectly to lib64. This led
to compatibility issues when running benchmarks.

This patch sets the correct path and adds a warning message so that it's
easier to notice such issues in the future.