[SYCL][E2E] Fix -Werror failures in DeviceLib tests
The `-fsycl-device-lib-jit-link` only has an effect in SPIR compilation
so these tests would fail with `-Werror`.

It might be worth re-working these tests a bit to only run this part on
spir, but other existing tests already have this change to make it work
on `-Werror` so do it as well here to fix the nightly.