[UR][Offload] Add initial membuffer implementation (#18849)
Add a basic implementation of buffers to the Offload adapter. The design
is loosely based on the CUDA and HIP adapters.

For now contexts only support a single device, so the implementation is
relatively simple. When multi-device contexts are supported we will need
to handle migration of data between devices (or otherwise change this
implementation)