[SYCL][E2E] Fix -Werror failures in implicitly SPIR only DeviceLib tests
The `-fsycl-device-lib-jit-link` only has an effect in SPIR compilation
so these tests would fail with `-Werror` for non SPIR targets.

This basically the same as #18985, but also apply it to tests that have
`UNSUPPORTED: target-amd || target-nvidia`. These tests could be enabled
for future non-SPIR targets we should not be assuming that
`!target-amd && !target-nvidia` implies SPIR.