[SYCL][E2E] Pass sycl target options to clang in one test
This is required to ensure that the correct target with the correct
options is used when linking with clang in the test.
For example without trying to use the `SPIR-V` (not the SPIR) backend
for testing with `--param spirv-backend=True` does not take effect.
If the test would be run on a non SPIR target it would fail
as clang is expecting object files for the SPIR target when `-fsycl-targets`
is not set.