[SYCL] Fix handling of discarded events in preview lib
The queue implementation assumed that handler.finalize()
will return either a valid (non-discarded) event or nullptr
(that's why parseEvent is nop in preview mode).

However, that was not always true. It was possible for
scheduler to only mark an event as discarded after
handler.finalize() completed (during actual command
execution). This resulted in discarded event being
stored in LastEventPtr in the queue and calls to wait()
on that discarded event (which is not allowed) in MT
scenarios.

Fix this, by modyfing addCG() to mark the event as
discarded immediately. This allows handler to return
nullptr for preview mode.