[SYCL][UR][L0 v2] Fix issues with event lifetime (#18962)
When releasing executionEvent in the commandBuffer we need to also
retain the queue. Otherwise event->release() in commandBuffer's
destructor might attempt to release the event to an event pool that has
already been destroyed.

Also, make sure eventPool is destroyed after commandListManager to avoid
any issues with events owned by the commandListManager and move context
retain/release from commandListManager to queue/command_buffer.

The implementation of commandListManager marked move ctor and assignment
operator as defaulted which was wrong - the implementation would have to
account for context and device retain/release calls. To avoid this, move
the logic to queue/commandBuffer which can have move ctors/assignments
operators removed.