[SYCL][E2E] Cleanup DeviceImageDependencies unsupported (#19016)
This feature is not supported on CUDA and HIP.

There was only one test without the unsupported directive, the
`free_function_kernels.cpp` test, but this is also tested in
`KernelAndProgram` without the image dependency flag and more
thoroughly. So since the feature is unsupported anyway it makes sense to
disable it as well.