[SYCL] Always store last event (for ioq) if scheduler was not bypassed (#18867)
This fixes an issue for cases where the kernel depends on events coming
from the scheduler (host task events or unenqueued commands). The logic
in `submit_impl` for calculating `noLastEventPath` did not take into
account all dependencies which could result in skipping storing last
even if `handler.finalize()` ended up submitting work to the scheduler.

I also restored the lock on the submit path. It is needed to store the
last event in case `handler.finalize()` returns an event coming from the
scheduler.