[AsyncAlloc][SYCL] Use the SYCL properties extension for memory pool creation (#17955)
Introduce initial_threshold, maximum_size, read_only, and zero_init
runtime properties conforming to the SYCL properties extension, and
apply these properties to the creation of memory pools.

This PR is a draft until the feature has been added to the
[sycl_ext_oneapi_async_memory_alloc](https://github.com/intel/llvm/pull/14800)
spec.

This is an ABI-break but on experimental extension which has been just recently added.

---------

Co-authored-by: omarahmed1111 <omar.ahmed@codeplay.com>