[SYCL] Fix NVPTX compilation with the new offload driver (#19039)
A typical SYCL compilation for NVPTX with the default
`nvptx(64)?-nvidia-cuda` triple compiles for an older SM_50
architecture, relying on forward compatibility and JIT to run on newer
devices. Thus compilation for NVPTX with the new offload driver relies
on us generating a fat binary containing the textual PTX as well as the
compiled object. This is a kind of LTO though not officially considered
so.

A recent pulldown broke the generation of textual assembly with the `-S`
flag, and so we were later passing a compiled ELF binary into `ptxas`
which predictably went wrong.

Fixes #18432.