fix native cpu and some tests (#19081)
We had this previously, but I refactored it upstream again and we lost
it.

Also I updated some tests that have `-fsycl-device-only` but no triple,
and we want that case to assert.

Also I fixed the stable name test to do the same check using native cpu,
and also fix it to actually check the `WIN:` prefix which it wasn't
doing before.

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>