[SYCL][Driver] Emit unused argument warning for -fno-libspirv
Previously the `-fno-libspirv` option was not warned about when there
was no SYCL compilation.
Remove the explicit target check for the targets that support
`-fno-libspirv` and instead rely on each target to emit appropriate
warnings when its used.
This commit slightly degrades diagnostic quality, from

"ignoring '-fno-sycl-libspirv' option as it is not currently supported for target"
to
"argument unused during compilation: '-fno-sycl-libspirv'"
, but I believe this is acceptable as it allows to remove the list
of targets that support the option from the driver. Additionally,
now if the user mixes targets that support and do not support
`-fno-libspirv` in the same compilation, they will not get warnings
that are not actionable.