[SYCL][Driver][libdevice] Link libspirv in `-fsycl-device-only` mode
Previously, the `-fsycl-device-only` mode disabled linking libspirv.

`-fsycl-device-only` should produce the same LLVM-IR as `-fsycl` would,
but without any host actions.

Libdevice is the likely reason why this was done as is, but it should
not be the default. This patch changes libdevice options to disable
libspirv linking there, and enables linking by default in
`-fsycl-device-only` mode.