[UR][HIP] Select the correct binary in a multi-arch bundle (#19092)
This commit changes the HIP adapter to select the correct binary for the
device when a bundle contains binaries built for multiple AMDGPU
architectures.

Similarly to other adapaters, the HIP adapter would previously select
the first 'amdgcn' binary it came across. This works fine for the common
case where the program was compiled for one architecture but may fail
otherwise.

To aid in this, the SYCL runtime passes some extra information into
urDeviceSelectBinary via the pre-existing 'pNext' field of
ur_device_binary_t. It does this only for the HIP backend. The HIP
adapater then parses this binary information as a clang offload bundle,
which conveniently contains specific triple & architecture information
for each binary. For this we re-use the code that the offload adapter
was using, making it common and fixing a bug in the version matching
logic.