[UR][L0] Init/deinit L0 adapter on first/last use (#19089)
Instead of using a global constructor to initialize the L0 adapter,
do it in the first call to `urAdapterGet`. Likewise, instead of
de-initing it as a global destructor, do it in the last call to
`urAdapterRelease`.

As well as not doing L0 initialization where the user is not using
L0, it also allows `urAdapterRelease` to be called in a global
destructor (e.g. what the SYCL runtime does) without worrying
about global destructor order.