[NFC][SYCL] Use `context_impl &` in `sampler_impl` ctor and near it
`SetArgBasedOnType` argument is only used to pass to the `sampler_impl`
ctor so update it.

`getCGKernelInfo` is only called in a function that also calls
`sampler_impl` ctor so updating its signuature allows to update that
caller's local `ContextImpl` variable, so makes sense to do as part of
this PR as well.

Continuation of the refactoring in
https://github.com/intel/llvm/pull/18795
https://github.com/intel/llvm/pull/18877
https://github.com/intel/llvm/pull/18966
https://github.com/intel/llvm/pull/18979
https://github.com/intel/llvm/pull/18980
https://github.com/intel/llvm/pull/18981
https://github.com/intel/llvm/pull/19007
https://github.com/intel/llvm/pull/19030
https://github.com/intel/llvm/pull/19123
https://github.com/intel/llvm/pull/19126