[NFC][SYCL] Move `platform::khr_get_default_context` impl to `platform_impl.cpp`
Noticed this code as part of the ongoing refactoring to prefer raw
ptrs/refs for SYCL RT `*_impl` objects. The usage of this API in
`queue_impl` doesn't need to go through user-visible SYCL objects, so
move the implementation and use the internal interfrace in `queue_impl`
and delegate `platform::khr_get_default_context` to `platform_impl`.