[CI][NFC] Remove Windows driver listing in dependencies.json (#19197)
We don't use this and it's all done manually, and JSON doesn't support
comments, so just remove it to prevent confusion.

The `tbb` and `oclcpu` ones aren't used either, but it's updated by the
team to have the download link we use to manually install, so keep
those.

Issue: https://github.com/intel/llvm/issues/19192

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>