[SYCL][CodeGen] Directly disable kernel stub for SYCL (NFC)
The way 1ce380335dc2a56f499afb44cba7b033c4118945
    and bdd5232e9b03cd912044a486dfb88ff9e5d8744e
disabled the kernel stub for SYCL is weird, returning from
`TargetCodeGenInfo::setOCLKernelStubCallingConvention()` in SYCL mode.

Basically we were actually emitting code for the stub would but we
customized it to not change its calling convention.

Instead, add `getLangOpts().OpenCL` checks to `CodeGenFunction::GenerateCode`
(to emit the kernel contents directly into the kernel) and to
`CodeGenModule::EmitGlobal` (to disable the deferred emission of the stub).
With this change, we should never come across `KernelReferenceKind::Stub`
in SYCL mode.